+++ kriegspiel/__init__.py
@@ -4,4 +4,4 @@ __author__ = "Alexander Filatov"
 
 __email__ = "alexander@kriegspiel.org"
 
-__version__ = "1.3.0"
+__version__ = "1.3.1"